Bugfix1 editor

+++ backend/src/middleware/imageURLProvider.js
@@ -1,5 +1,6 @@
 const fs = require("fs")
 const path = require("path")
+const models = require("../models")
 
 const imageURLProvider = (req, res) => {
   //   const imageUrl = `${req.protocol}://${req.get("host")}/src/images/${
@@ -21,14 +22,23 @@ const deleteImageForm = (req, res) => {
   // Construire le chemin d'accès absolu au fichier
   const absolutePath = path.join(__dirname, "..", "..", relativePath)
 
-  // Supprimer le fichier
-  fs.unlink(absolutePath, (err) => {
-    if (err) {
-      console.error(err)
-    } else {
-      console.info("Image supprimée avec succès")
-    }
-  })
+  models.campagnes
+    .verifyNumberOfSameImageBetweenCampagnesAndScenarios(absolutePath)
+    .then(([result]) => {
+      const countImage = result[0].countImage
+      if (countImage === 1) {
+        // Supprimer le fichier
+        fs.unlink(absolutePath, (err) => {
+          if (err) {
+            console.error(err)
+          } else {
+            console.info("Image supprimée avec succès")
+          }
+        })
+      } else {
+        console.info("Image non supprimée car existante dans une autre table")
+      }
+    })
 }
 
 module.exports = {
